v3.0

+++ customize.sh
@@ -9,6 +9,22 @@ if [ "$BOOTMODE" != true ]; then
   ui_print " "
 fi
 
+# optionals
+OPTIONALS=/sdcard/optionals.prop
+if [ ! -f $OPTIONALS ]; then
+  touch $OPTIONALS
+fi
+
+# debug
+if [ "`grep_prop debug.log $OPTIONALS`" == 1 ]; then
+  ui_print "- The install log will contain detailed information"
+  set -x
+  ui_print " "
+fi
+
+# var
+LIST32BIT=`getprop ro.product.cpu.abilist32`
+
 # run
 . $MODPATH/function.sh
 
@@ -94,12 +110,6 @@ SYSTEM_EXT=`realpath $MIRROR/system_ext`
 ODM=`realpath $MIRROR/odm`
 MY_PRODUCT=`realpath $MIRROR/my_product`
 
-# optionals
-OPTIONALS=/sdcard/optionals.prop
-if [ ! -f $OPTIONALS ]; then
-  touch $OPTIONALS
-fi
-
 # .aml.sh
 mv -f $MODPATH/aml.sh $MODPATH/.aml.sh
 
@@ -112,53 +122,44 @@ if [ ! -e /dev/socket/audio_hw_socket ]; then
 fi
 
 # function
-check_function() {
+run_check_function() {
+LISTS=`strings $MODPATH/system/vendor$DIR/$DES | grep ^lib | grep .so`
+FILE=`for LIST in $LISTS; do echo $SYSTEM$DIR/$LIST; done`
 ui_print "- Checking"
 ui_print "$NAME"
 ui_print "  function at"
 ui_print "$FILE"
 ui_print "  Please wait..."
 if ! grep -q $NAME $FILE; then
-  ui_print "  ! Function not found."
-  ui_print "    Unsupported Dolby Atmos."
-  DOLBY=false
-else
-  DOLBY=true
+  ui_print "  Function not found."
+  ui_print "  Using new $DIR$LIB"
+  mv -f $MODPATH/system_support$DIR/$LIB $MODPATH/system$DIR
 fi
 ui_print " "
 }
-check_function_2() {
+check_function() {
 if [ "$IS64BIT" == true ]; then
-  LISTS=`strings $MODPATH/system_dolby/vendor/lib64/$DES | grep ^lib | grep .so`
-  FILE=`for LIST in $LISTS; do echo $SYSTEM/lib64/$LIST; done`
-  ui_print "- Checking"
-  ui_print "$NAME"
-  ui_print "  function at"
-  ui_print "$FILE"
-  ui_print "  Please wait..."
-  if ! grep -q $NAME $FILE; then
-    ui_print "  Using new $LIB 64"
-    mv -f $MODPATH/system_support/lib64/$LIB $MODPATH/system/lib64
-  fi
-  ui_print " "
+  DIR=/lib64
+  run_check_function
 fi
-LISTS=`strings $MODPATH/system_dolby/vendor/lib/$DES | grep ^lib | grep .so`
-FILE=`for LIST in $LISTS; do echo $SYSTEM/lib/$LIST; done`
-ui_print "- Checking"
-ui_print "$NAME"
-ui_print "  function at"
-ui_print "$FILE"
-ui_print "  Please wait..."
-if ! grep -q $NAME $FILE; then
-  ui_print "  Using new $LIB"
-  mv -f $MODPATH/system_support/lib/$LIB $MODPATH/system/lib
+if [ "$LIST32BIT" ] && [ $DOLBY32 == true ]; then
+  DIR=/lib
+  run_check_function
 fi
-ui_print " "
 }
 
 # bit
 if [ "$IS64BIT" == true ]; then
-  ui_print "- 64 bit"
+  ui_print "- 64 bit architecture"
+  ui_print " "
+  # 32 bit
+  if [ "$LIST32BIT" ]; then
+    ui_print "- 32 bit library support"
+  else
+    ui_print "- Doesn't support 32 bit library"
+    rm -rf $MODPATH/armeabi-v7a $MODPATH/x86\
+     $MODPATH/system*/lib $MODPATH/system*/vendor/lib
+  fi
   ui_print " "
   # check
   NAME=_ZN7android23sp_report_stack_pointerEv
@@ -166,18 +167,41 @@ if [ "$IS64BIT" == true ]; then
     ui_print "- Activating Dolby Atmos..."
     ui_print " "
     FILE=$VENDOR/lib64/hw/*audio*.so
-    check_function
+    ui_print "- Checking"
+    ui_print "$NAME"
+    ui_print "  function at"
+    ui_print "$FILE"
+    ui_print "  Please wait..."
+    if grep -q $NAME $FILE; then
+      DOLBY=true
+    else
+      ui_print "  ! Function not found."
+      ui_print "    Unsupported Dolby Atmos."
+      DOLBY=false
+    fi
   fi
-  if [ $DOLBY == true ]; then
+  if [ $DOLBY == true ] && [ "$LIST32BIT" ]; then
     FILE=$VENDOR/lib/hw/*audio*.so
-    check_function
+    ui_print "- Checking"
+    ui_print "$NAME"
+    ui_print "  function at"
+    ui_print "$FILE"
+    ui_print "  Please wait..."
+    if grep -q $NAME $FILE; then
+      DOLBY32=true
+    else
+      ui_print "  Function not found."
+      DOLBY32=false
+      rm -rf $MODPATH/system_dolby/lib $MODPATH/system_dolby/vendor/lib
+    fi
+    ui_print " "
   fi
   # check
   NAME=_ZN7android8hardware23getOrCreateCachedBinderEPNS_4hidl4base4V1_05IBaseE
   DES=vendor.dolby.hardware.dms@2.0.so
   LIB=libhidlbase.so
   if [ $DOLBY == true ]; then
-    check_function_2
+    check_function
   fi
   if [ $DOLBY == true ]; then
     MODNAME2='Mi Sound and Dolby Atmos Redmi K40'
@@ -188,7 +212,7 @@ if [ "$IS64BIT" == true ]; then
     cp -rf $MODPATH/system_dolby/* $MODPATH/system
   fi
 else
-  ui_print "- 32 bit"
+  ui_print "- 32 bit architecture"
   rm -rf `find $MODPATH -type d -name *64*`
   if [ $DOLBY == true ]; then
     ui_print "  ! Unsupported Dolby Atmos."
@@ -206,45 +230,55 @@ FILE4=$ODM/lib64/soundfx/libmisoundfx.so
 if [ -f $FILE ] || [ -f $FILE2 ] || [ -f $FILE3 ]\
 || [ -f $FILE4 ]; then
   ui_print "- Built-in misoundfx is detected."
-  BUILTIN=true
   rm -f `find $MODPATH/system/vendor -type f -name *misound*`
   ui_print " "
-fi
-
-# function
-legacy_libraries() {
-ui_print "  Using legacy libraries."
-cp -rf $MODPATH/system_10/* $MODPATH/system
-rm -f $MODPATH/system/vendor/lib64/soundfx/libmisoundfx.so
-if [ "$BUILTIN" != true ]\
-&& [ ! "`getprop ro.product.cpu.abilist32`" ]; then
-  abort "  This ROM doesn't support 32 bit library."
-fi
-}
-
-# check
-NAME=_ZN7android23sp_report_stack_pointerEv
-FILE=$VENDOR/lib/hw/*audio*.so
-FILE2=$VENDOR/lib64/hw/*audio*.so
-ui_print "- Checking"
-ui_print "$NAME"
-ui_print "  function at"
-ui_print "$FILE"
-if [ "$IS64BIT" == true ]; then
-  ui_print "$FILE2"
-fi
-ui_print "  Please wait..."
-if [ "$IS64BIT" == true ]; then
-  if ! grep -q $NAME $FILE\
-  || ! grep -q $NAME $FILE2; then
-    legacy_libraries
-  fi
 else
-  if ! grep -q $NAME $FILE; then
-    legacy_libraries
+  NAME=_ZN7android23sp_report_stack_pointerEv
+  if [ "$IS64BIT" == true ]; then
+    FILE=$VENDOR/lib64/hw/*audio*.so
+    ui_print "- Checking"
+    ui_print "$NAME"
+    ui_print "  function at"
+    ui_print "$FILE"
+    ui_print "  Please wait..."
+    if grep -q $NAME $FILE; then
+      FUNC64=true
+    else
+      ui_print "  Function not found."
+      FUNC64=false
+    fi
+  else
+    FUNC64=false
+  fi
+  if [ "$LIST32BIT" ]; then
+    FILE=$VENDOR/lib/hw/*audio*.so
+    ui_print "- Checking"
+    ui_print "$NAME"
+    ui_print "  function at"
+    ui_print "$FILE"
+    ui_print "  Please wait..."
+    if grep -q $NAME $FILE; then
+      FUNC32=true
+    else
+      ui_print "  Function not found."
+      FUNC32=false
+    fi
+  else
+    FUNC32=false
+  fi
+  if [ $FUNC64 != true ] && [ $FUNC32 != true ]; then
+    if [ ! "$LIST32BIT" ]; then
+      abort "  ! This ROM doesn't support 32 bit library."
+    fi
+    ui_print "  Using legacy libraries."
+    cp -rf $MODPATH/system_10/* $MODPATH/system
+    rm -f $MODPATH/system/vendor/lib64/soundfx/libmisoundfx.so
+  elif [ $FUNC64 == true ] && [ $FUNC32 != true ]; then
+    rm -f $MODPATH/system/vendor/lib/soundfx/libmisoundfx.so
+  elif [ $FUNC64 != true ] && [ $FUNC32 == true ]; then
+    rm -f $MODPATH/system/vendor/lib64/soundfx/libmisoundfx.so
   fi
 fi
-ui_print " "
 rm -rf $MODPATH/system_10
 
 # sepolicy
@@ -469,15 +503,14 @@ early_init_mount_dir() {
 if echo $MAGISK_VER | grep -q delta\
 && [ "`grep_prop dolby.skip.early $OPTIONALS`" != 1 ]; then
   EIM=true
-  DIR=$MAGISKTMP/mirror/early-mount
   if "$BOOTMODE"\
-  && [ -L $DIR ]; then
-    EIMDIR=`readlink $DIR`
-    [ "${EIMDIR:0:1}" != "/" ] && EIMDIR="$MAGISKTMP/mirror/$EIMDIR"
+  && [ -L $MIRROR/early-mount ]; then
+    EIMDIR=`readlink $MIRROR/early-mount`
+    [ "${EIMDIR:0:1}" != "/" ] && EIMDIR="$MIRROR/$EIMDIR"
   elif "$BOOTMODE"\
-  && [ "$MAGISK_VER_CODE" -ge 26000 ]; then
-    DIR=$MAGISKTMP/preinit
-    MOUNT=`mount | grep $DIR`
+  && [ "$MAGISK_VER_CODE" -ge 26000 ]\
+  && [ -d $MAGISKTMP/preinit ]; then
+    MOUNT=`mount | grep $MAGISKTMP/preinit`
     BLOCK=`echo $MOUNT | sed 's| on.*||g'`
     DIR=`mount | sed "s|$MOUNT||g" | grep -m 1 $BLOCK`
     EIMDIR=`echo $DIR | sed "s|$BLOCK on ||g" | sed 's| type.*||g'`/early-mount.d
@@ -537,63 +570,45 @@ else
   EIM=false
 fi
 }
-find_file() {
+run_find_file() {
 for NAME in $NAMES; do
-  if [ "$IS64BIT" == true ]; then
-    FILE=`find $SYSTEM/lib64 $VENDOR/lib64 $SYSTEM_EXT/lib64 -type f -name $NAME`
-    if [ ! "$FILE" ]; then
-      if [ "`grep_prop install.hwlib $OPTIONALS`" == 1 ]; then
-        ui_print "- Installing $NAME 64 directly to"
-        ui_print "$SYSTEM/lib64..."
-        cp $MODPATH/system_support/lib64/$NAME $SYSTEM/lib64
-        DES=$SYSTEM/lib64/$NAME
-        if [ -f $MODPATH/system_support/lib64/$NAME ]\
-        && [ ! -f $DES ]; then
-          ui_print "  ! Installation failed."
-          ui_print "    Using $NAME 64 systemlessly."
-          cp -f $MODPATH/system_support/lib64/$NAME $MODPATH/system/lib64
-        fi
-      else
-        ui_print "! $NAME 64 not found."
-        ui_print "  Using $NAME 64 systemlessly."
-        cp -f $MODPATH/system_support/lib64/$NAME $MODPATH/system/lib64
-        ui_print "  If this module still doesn't work, type:"
-        ui_print "  install.hwlib=1"
-        ui_print "  inside $OPTIONALS"
-        ui_print "  and reinstall this module"
-        ui_print "  to install $NAME 64 directly to this ROM."
-        ui_print "  DwYOR!"
-      fi
-      ui_print " "
-    fi
-  fi
-  FILE=`find $SYSTEM/lib $VENDOR/lib $SYSTEM_EXT/lib -type f -name $NAME`
+  FILE=`find $SYSTEM$DIR $SYSTEM_EXT$DIR -type f -name $NAME`
   if [ ! "$FILE" ]; then
     if [ "`grep_prop install.hwlib $OPTIONALS`" == 1 ]; then
-      ui_print "- Installing $NAME directly to"
-      ui_print "$SYSTEM/lib..."
-      cp $MODPATH/system_support/lib/$NAME $SYSTEM/lib
-      DES=$SYSTEM/lib/$NAME
-      if [ -f $MODPATH/system_support/lib/$NAME ]\
+      ui_print "- Installing $DIR/$NAME directly to"
+      ui_print "$SYSTEM..."
+      cp $MODPATH/system_support$DIR/$NAME $SYSTEM$DIR
+      DES=$SYSTEM$DIR/$NAME
+      if [ -f $MODPATH/system_support$DIR/$NAME ]\
       && [ ! -f $DES ]; then
         ui_print "  ! Installation failed."
-        ui_print "    Using $NAME systemlessly."
-        cp -f $MODPATH/system_support/lib/$NAME $MODPATH/system/lib
+        ui_print "    Using $DIR/$NAME systemlessly."
+        cp -f $MODPATH/system_support$DIR/$NAME $MODPATH/system$DIR
       fi
     else
-      ui_print "! $NAME not found."
-      ui_print "  Using $NAME systemlessly."
-      cp -f $MODPATH/system_support/lib/$NAME $MODPATH/system/lib
+      ui_print "! $DIR/$NAME not found."
+      ui_print "  Using $DIR/$NAME systemlessly."
+      cp -f $MODPATH/system_support$DIR/$NAME $MODPATH/system$DIR
       ui_print "  If this module still doesn't work, type:"
       ui_print "  install.hwlib=1"
       ui_print "  inside $OPTIONALS"
       ui_print "  and reinstall this module"
-      ui_print "  to install $NAME directly to this ROM."
+      ui_print "  to install $DIR/$NAME directly to this ROM."
       ui_print "  DwYOR!"
     fi
     ui_print " "
   fi
 done
+}
+find_file() {
+if [ "$IS64BIT" == true ]; then
+  DIR=/lib64
+  run_find_file
+fi
+if [ "$LIST32BIT" ]; then
+  DIR=/lib
+  run_find_file
+fi
 sed -i 's|^install.hwlib=1|install.hwlib=0|g' $OPTIONALS
 }
 patch_manifest_eim() {
@@ -1009,24 +1024,29 @@ done
 }
 
 # check
-FILES=/lib/libmigui.so
-file_check_system
 if "$IS64BIT"; then
   FILES=/lib64/libmigui.so
   file_check_system
 fi
+if [ "LIST32BIT" ]; then
+  FILES=/lib/libmigui.so
+  file_check_system
+fi
 if [ $DOLBY == true ]; then
-  FILES="/etc/acdbdata/adsp_avs_config.acdb
-         /lib/libstagefrightdolby.so
-         /lib/libdeccfg.so
-         /lib/libstagefright_soft_ddpdec.so
-         /lib/libstagefright_soft_ac4dec.so"
+  FILES=/etc/acdbdata/adsp_avs_config.acdb
   file_check_vendor
   if "$IS64BIT"; then
-    FILES="/lib64/libstagefrightdolby.so
-           /lib64/libdeccfg.so
+    FILES="/lib64/libdeccfg.so
+           /lib64/libstagefrightdolby.so
            /lib64/libstagefright_soft_ddpdec.so
            /lib64/libstagefright_soft_ac4dec.so"
+    file_check_vendor
+  fi
+  if [ "LIST32BIT" ]; then
+    FILES="/lib/libdeccfg.so
+           /lib/libstagefrightdolby.so
+           /lib/libstagefright_soft_ddpdec.so
+           /lib/libstagefright_soft_ac4dec.so"
     file_check_vendor
   fi
 fi
